backend/requirements.txt
djangorestframework==3.14.0
django-cors-headers==4.1.0
